MHD-Kontrolle: move shelf-life warning flag to picking profile + reword message (#24921)
* MHD-Kontrolle — move IsWarnShelfLifeUndercut from C_Workplace to picking profile (AD)

* MHD-Kontrolle — backup_table before C_Workplace DROP COLUMN (review)

* MHD-Kontrolle — regenerate models for flag move (MobileUI_UserProfile_Picking + C_Workplace)

* MHD-Kontrolle — drop C_Workplace column without backup (just-added + unused, per decision)

* MHD-Kontrolle — read shelf-life flag from picking profile; revert Workplace wiring

* MHD-Kontrolle — review fixes: rename isWarnShelfLifeUndercut, hoist getPickingJobOptions, drop internal refs

* MHD-Kontrolle — reword RLZ_TooShort warning to spell out Restlaufzeit + question form

* MHD-Kontrolle — set shelf-life flag via mobile picking profile in test masterdata + E2E

* MHD-Kontrolle — review: drop transitional absence-marker comments in WorkplaceRepository

* MHD-Kontrolle — desktop E2E: shelf-life checkbox on the picking-profile window (replaces stale C_Workplace test)
