+++ backend/de.metas.adempiere.adempiere/base/src/main/java-gen/org/compiere/model/I_MobileUI_UserProfile_Picking.java
@@ -598,6 +598,27 @@ public interface I_MobileUI_UserProfile_Picking
 	String COLUMNNAME_IsShowLastPickedBestBeforeDateForLines = "IsShowLastPickedBestBeforeDateForLines";
 
 	/**
+	 * Set Warn when guaranteed remaining shelf life is undercut.
+	 *
+	 * <br>Type: YesNo
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	void setIsWarnShelfLifeUndercut (boolean IsWarnShelfLifeUndercut);
+
+	/**
+	 * Get Warn when guaranteed remaining shelf life is undercut.
+	 *
+	 * <br>Type: YesNo
+	 * <br>Mandatory: true
+	 * <br>Virtual Column: false
+	 */
+	boolean isWarnShelfLifeUndercut();
+
+	ModelColumn<I_MobileUI_UserProfile_Picking, Object> COLUMN_IsWarnShelfLifeUndercut = new ModelColumn<>(I_MobileUI_UserProfile_Picking.class, "IsWarnShelfLifeUndercut", null);
+	String COLUMNNAME_IsWarnShelfLifeUndercut = "IsWarnShelfLifeUndercut";
+
+	/**
 	 * Set Lot number.
 	 *
 	 * <br>Type: YesNo
