+++ backend/de.metas.adempiere.adempiere/base/src/main/java-gen/org/compiere/model/X_MobileUI_UserProfile_Picking.java
@@ -12,7 +12,7 @@ import javax.annotation.Nullable;
 public class X_MobileUI_UserProfile_Picking extends org.compiere.model.PO implements I_MobileUI_UserProfile_Picking, org.compiere.model.I_Persistent 
 {
 
-	private static final long serialVersionUID = 884988273L;
+	private static final long serialVersionUID = -1470502703L;
 
     /** Standard Constructor */
     public X_MobileUI_UserProfile_Picking (final Properties ctx, final int MobileUI_UserProfile_Picking_ID, @Nullable final String trxName)
@@ -321,6 +321,18 @@ public class X_MobileUI_UserProfile_Picking extends org.compiere.model.PO implem
 	public boolean isShowLastPickedBestBeforeDateForLines() 
 	{
 		return get_ValueAsBoolean(COLUMNNAME_IsShowLastPickedBestBeforeDateForLines);
+	}
+
+	@Override
+	public void setIsWarnShelfLifeUndercut (final boolean IsWarnShelfLifeUndercut)
+	{
+		set_Value (COLUMNNAME_IsWarnShelfLifeUndercut, IsWarnShelfLifeUndercut);
+	}
+
+	@Override
+	public boolean isWarnShelfLifeUndercut() 
+	{
+		return get_ValueAsBoolean(COLUMNNAME_IsWarnShelfLifeUndercut);
 	}
 
 	@Override
