+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/workplace/WorkplaceCreateRequest.java
@@ -28,7 +28,6 @@ public class WorkplaceCreateRequest
 	@Nullable SeqNo seqNo;
 	@Nullable OrderPickingType orderPickingType;
 	int maxPickingJobs;
-	boolean warnShelfLifeUndercut;
 
 	@NonNull @Singular ImmutableSet<ProductCategoryId> productCategoryIds;
 	@NonNull @Singular ImmutableSet<ProductId> productIds;
