+++ backend/de.metas.adempiere.adempiere/base/src/test/java/de/metas/workplace/WorkplaceRepositoryTest.java
@@ -27,9 +27,7 @@ import org.adempiere.warehouse.WarehouseId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class WorkplaceRepositoryTest
 {
@@ -55,23 +53,4 @@ public class WorkplaceRepositoryTest
 		assertNotNull(workplace2);
 	}
 
-	@Test
-	public void isWarnShelfLifeUndercut_roundTrip()
-	{
-		final WorkplaceRepository repo = WorkplaceRepository.newInstanceForUnitTesting();
-
-		final Workplace withFlagTrue = repo.create(WorkplaceCreateRequest.builder()
-				.name("WarnTrue")
-				.warehouseId(WarehouseId.ofRepoId(1))
-				.warnShelfLifeUndercut(true)
-				.build());
-		assertTrue(withFlagTrue.isWarnShelfLifeUndercut(), "Workplace created with warnShelfLifeUndercut=true must load back as true");
-
-		final Workplace withFlagFalse = repo.create(WorkplaceCreateRequest.builder()
-				.name("WarnFalse")
-				.warehouseId(WarehouseId.ofRepoId(1))
-				.warnShelfLifeUndercut(false)
-				.build());
-		assertFalse(withFlagFalse.isWarnShelfLifeUndercut(), "Workplace created with warnShelfLifeUndercut=false must load back as false");
-	}
 }
