+++ backend/de.metas.frontend-testing/src/main/java/de/metas/frontend_testing/masterdata/mobile_configuration/JsonMobileConfigRequest.java
@@ -60,6 +60,7 @@ public class JsonMobileConfigRequest
 		@Nullable Boolean considerOnlyJobScheduledToWorkplace;
 		@Nullable Boolean allowQuickPackAll;
 		@Nullable Boolean showPromptWhenOverPicking;
+		@Nullable Boolean warnShelfLifeUndercut;
 
 		@Nullable List<Customer> customers;
