+++ backend/de.metas.frontend-testing/src/main/java/de/metas/frontend_testing/masterdata/mobile_configuration/MobileConfigPickingCommand.java
@@ -149,6 +149,11 @@ class MobileConfigPickingCommand
 
 		builder.isShowConfirmationPromptWhenOverPick(Boolean.TRUE.equals(from.getShowPromptWhenOverPicking()));
 
+		if (from.getWarnShelfLifeUndercut() != null)
+		{
+			builder.isWarnShelfLifeUndercut(from.getWarnShelfLifeUndercut());
+		}
+
 		return builder.build();
 	}
