+++ backend/de.metas.frontend-testing/src/main/java/de/metas/frontend_testing/masterdata/workplace/CreateWorkplaceCommand.java
@@ -56,7 +56,6 @@ public class CreateWorkplaceCommand
 						.pickingSlotId(request.getPickingSlot() != null
 								? context.getId(request.getPickingSlot(), PickingSlotId.class)
 								: null)
-						.warnShelfLifeUndercut(Boolean.TRUE.equals(request.getWarnShelfLifeUndercut()))
 						.build()
 		);
 		context.putIdentifier(identifier, workplace.getId());
