+++ backend/de.metas.frontend-testing/src/main/java/de/metas/frontend_testing/masterdata/workplace/JsonWorkplaceRequest.java
@@ -15,6 +15,4 @@ public class JsonWorkplaceRequest
 	@Nullable Identifier warehouse;
 	@Nullable Identifier pickingSlot;
 	@Nullable Identifier pickFromLocator;
-	/** C_Workplace.IsWarnShelfLifeUndercut — when true, a shelf-life undercut warning dialog is shown to the picker */
-	@Nullable Boolean warnShelfLifeUndercut;
 }
